Initial working commit
* Add README.md

* Add CircleCI configuration

* Read the configuration from a external JSON file

* Store the application key at compile time
Compile the `spotify_key.c` file that you get from Spotify into a small
library that we later link against.
This way we can separate the real application key that is used in the
release version and one that might be used in development, meaning a
private key for each developer.