Add playback toggling
When listening to music you probably want to be able to pause the
playback for whatever reason.

This implements a action for toggling the playback status and also binds
that to `Space`. I also added a label to the status bar to display the
current playback status.

This however does no longer build under Rust stable.